Ignore mouse positions that are outside the game
Depending on how the events are dispatched, the mouse may be recorded
as being outside the game if the game window is resized to a smaller
size. This means our hovering logic will not return a tile as the mouse
is not currently in the game window (and thus not over any tile).

When this happens, ignore such mouse positions and instead only process
positions where the mouse is over and inside the game window.

Signed-off-by: Remy Suen <remy.suen@gmail.com>