Always reset mouse state when button is released
Currently, the game only registers the mouse button being released if it
is released within the bounds of the game. This means that the game will
think that the mouse button is down even when it is not if the button
was released outside the game's bounds.

Instead, the game should always evaluate and update its internal state
when the mouse button is released. This should prevent inconsistencies
in the game's cached state of the mouse and the actual physical state of
the mouse.

Signed-off-by: Remy Suen <remy.suen@gmail.com>